Plot coefficient paths for multiclass regression objects

Plot the coefficient paths.

# S3 method for mcreg
plot(
  obj,
  class_var = "class",
  x_vars = NULL,
  coef_type = c("rescaled", "scaled"),
  plot_type = c("knot", "lambda"),
  log_x = NULL
)

Arguments

obj
an multiclass regression object (object of class mcreg)

class_var
the name of the class variable in the data frame

y_var
the name of the dependent variable in the data frame

...
additional arguments to be passed on to multiclass_reg()

Value

A list with the following components

fit

fitted genlasso object

coef

matrix of estimated coefficients

lambda

vector of regularization parameters along knots of regularization path

K

number of classes

p

number of predictors per class

n

vector of sample sizes for each class

Y

composite Y vector (stacked over the classes)

X

block diagonal X matrix (each block coresponds to a class)

var_indicator

matrix of variable indicators

Examples

p = 7
k = 2
n = 20
beta = c(1,2)
set.seed(1)
X = list(matrix(rnorm(p*n), ncol = p), matrix(rnorm(p*n), ncol = p))
Y = list(rnorm(n), rnorm(n))
df = lists_to_data(Y, X)
mc_reg(df)

#> $fit
#> 
#> Call:
#> genlasso::genlasso(y = get.input$Y, X = get.input$X, D = get.input$D)
#> 
#> Output:
#> Path model with 7 total steps.
#> 
#> 
#> $coef
#>          4.342       3.924       3.716       2.569        1.51       0.938
#> V1 -0.06392407 -0.05991624 -0.05731597 -0.04975394  0.03931913  0.08455441
#> V2 -0.06551148 -0.06375715 -0.06221844 -0.11167921 -0.14291429 -0.16456523
#> V3  0.12744054  0.12579262  0.13536632  0.17106660  0.17530981  0.18084474
#> V4  0.25699921  0.26494024  0.26698565  0.28350161  0.31050907  0.32926742
#> V5 -0.07675868 -0.04755220 -0.03525306  0.02770356  0.11660557  0.16568744
#> V6 -0.10297332 -0.10852131 -0.11222892 -0.12892920 -0.14870904 -0.19326974
#> V7  0.24428787  0.24532396  0.24514843  0.24578718  0.23927103  0.23762769
#> V1 -0.06392407 -0.05991624 -0.05731597 -0.04975394 -0.09458315 -0.10807406
#> V2 -0.06551148 -0.06375715 -0.06221844 -0.00920820  0.04183691  0.07946326
#> V3  0.12744054  0.12579262  0.11615371  0.07691214  0.06973130  0.06937525
#> V4  0.25699921  0.26494024  0.26698565  0.28350161  0.31050907  0.32926742
#> V5 -0.07675868 -0.09979430 -0.10949299 -0.17344932 -0.24369766 -0.27156427
#> V6 -0.10297332 -0.10852131 -0.11222892 -0.12892920 -0.14870904 -0.12535171
#> V7  0.24428787  0.24532396  0.24514843  0.24578718  0.23927103  0.23762769
#>          0.817           0
#> V1  0.09698108  0.22127287
#> V2 -0.16743044 -0.20739002
#> V3  0.18011501  0.13086070
#> V4  0.33394737  0.46686326
#> V5  0.17717940  0.30174199
#> V6 -0.19951555 -0.26645668
#> V7  0.22725429  0.15960253
#> V1 -0.11050520 -0.12680234
#> V2  0.08700538  0.14202073
#> V3  0.06955799  0.06027143
#> V4  0.33394737  0.29771243
#> V5 -0.27734113 -0.29852069
#> V6 -0.12131008 -0.08779596
#> V7  0.24052540  0.24046893
#> 
#> $lambda
#> [1] 4.342 3.924 3.716 2.569 1.510 0.938 0.817 0.000
#> 
#> $K
#> [1] 2
#> 
#> $p
#> [1] 7
#> 
#> $n
#>  1  2 
#> 20 20 
#> 
#> $Y_list
#> $Y_list$`1`
#>               y
#> 1   0.256670132
#> 2  -0.548988603
#> 3   1.182798606
#> 4  -0.842691693
#> 5  -0.722074328
#> 6  -1.142532205
#> 7  -0.809638631
#> 8   0.803724929
#> 9   0.292242126
#> 10  0.863699194
#> 11 -0.531578688
#> 12  0.234910268
#> 13  0.102704901
#> 14 -1.567717366
#> 15  1.636969264
#> 16 -0.007012363
#> 17  0.624138975
#> 18  0.813676653
#> 19 -0.192025725
#> 20 -0.447275444
#> 
#> $Y_list$`2`
#>             y
#> 21  0.6831596
#> 22 -1.2578123
#> 23  1.7608233
#> 24 -0.5941462
#> 25  1.4436312
#> 26  1.3016986
#> 27 -0.1275484
#> 28  0.3567068
#> 29 -1.2350626
#> 30  0.1124924
#> 31  0.8330983
#> 32 -0.1114356
#> 33 -0.6646510
#> 34 -0.8662960
#> 35 -0.2464365
#> 36  0.8586473
#> 37 -0.6944891
#> 38 -0.3315242
#> 39 -1.5046541
#> 40  0.2837987
#> 
#> 
#> $X_list
#> $X_list$`1`
#>             V1         V2         V3           V4          V5          V6
#> 1  -0.47876585  0.7212135 -0.1765883  2.052937211 -0.44724339 -0.78208645
#> 2   0.26714948  0.5764173 -0.2796235 -0.206029892 -0.01580652 -0.03067421
#> 3  -0.67136830 -0.1722877  0.8225721  0.468626546  1.29123874 -1.11164466
#> 4   1.56694673 -2.3561876  0.6598507 -0.143798560 -1.39761854  0.10079849
#> 5   0.40145748  0.4046711 -0.7845974 -0.857599138  0.70986474 -0.82089774
#> 6  -0.65740917 -0.3105792 -0.8063316  0.005009635  0.45010515  1.92607908
#> 7   0.54686708 -0.4160399  0.4370732 -1.840168657  1.17679622  0.73447257
#> 8   0.77788495 -1.8074402  0.9055791  1.186628929 -0.18401133  0.95390972
#> 9   0.62821947 -0.7571342 -0.1160725 -0.027880993  0.48699801  0.35731367
#> 10 -0.18313757  0.1910506  1.0361444  1.840996461  0.38456548  1.82954281
#> 11  1.49006255  1.1864772  0.4759545  0.270360939 -0.42121855 -0.79951942
#> 12  0.45701273 -0.3599506 -0.6956061 -0.826755844  1.32087849 -0.60205814
#> 13 -0.47396565  0.1590209  0.4098614  0.395501520  1.27363812  1.54610396
#> 14 -1.94117948 -0.3081204 -1.2956179 -1.034203486  0.81562886 -0.81648750
#> 15  1.13386140 -1.7083006  1.6762616 -1.329927719  1.69804915 -0.31366251
#> 16  0.05668208 -0.6903072  2.3111135  0.100013886  0.67457294 -0.52398089
#> 17  0.08314817 -0.6683990 -0.4116791 -0.579972159 -1.15181375 -0.44139136
#> 18  0.96711434 -0.3139490 -1.1967691 -0.168691040 -0.45181830 -0.39502424
#> 19  0.85421378  0.9127865  0.6749937 -0.100912550 -1.10008037  0.48208311
#> 20  0.64490384  0.5563545 -0.1424099 -0.715304364 -0.35242654 -0.27957839
#>              V7
#> 1  -0.463787962
#> 2   1.506485622
#> 3  -0.153298165
#> 4  -0.115978105
#> 5  -0.031406644
#> 6   0.834766414
#> 7  -0.003033862
#> 8   0.035253066
#> 9  -0.651015494
#> 10 -0.270183749
#> 11  0.139462113
#> 12 -0.552164899
#> 13  0.641713193
#> 14 -1.542631192
#> 15  0.402020957
#> 16 -1.561871226
#> 17 -0.245361720
#> 18 -0.487574553
#> 19 -0.619510556
#> 20  0.014727009
#> 
#> $X_list$`2`
#>             V1          V2         V3         V4          V5          V6
#> 21 -1.66463418  0.19862558 -1.4138724  0.2091795 -1.60627712  0.72381436
#> 22  1.18142151 -0.50370801  1.1553587  1.3933059  0.12085357  1.09447970
#> 23 -1.43500545  0.86882915  0.2692985  1.2986432 -0.50858307  0.17503617
#> 24 -0.32875042  0.68676624 -1.6875022 -0.4759629 -0.22062175 -1.07510631
#> 25 -0.92945305 -0.90642941  0.6185146 -2.2846557 -1.03236239  1.24063387
#> 26 -0.59327792  2.08316395 -0.1698973  2.1418239  1.91333305 -2.34710475
#> 27  2.01986170 -0.52104016  1.7128693  0.4476429 -0.21083180 -0.69636543
#> 28  0.11407313 -1.75849329 -0.8742812  0.3312741 -1.47917701 -0.36843478
#> 29 -1.08633470 -0.40398155 -0.4847358 -0.1821150  0.31499123 -0.26686423
#> 30 -1.41256907 -0.03158415 -1.0598825  0.3023816  0.38066091  1.07900436
#> 31  0.51257571  2.19096231 -0.1911791 -0.3218407 -0.86242547  0.07606432
#> 32  0.08096633 -0.13923441  0.4804846  0.2196308 -2.75650534  0.38338098
#> 33 -0.19481296  0.23237838 -0.8344605 -0.5401347 -0.51871622 -0.15744866
#> 34 -0.75767558 -0.33282558  0.9773020 -1.4378152  0.68653721 -0.35935366
#> 35 -1.27155713 -0.60460502 -1.3869177  0.7445114  0.05929142  0.71047589
#> 36 -0.89195200 -0.28793224 -1.2012341  1.2367809  0.02101810  1.22165106
#> 37  1.01885494  0.58224057  1.6856955 -0.4554475  0.67689618 -2.80412432
#> 38 -0.47398969  1.94469926 -1.1639614 -1.3296097 -1.06210711  0.57117882
#> 39 -1.17668673  0.83593112  0.4920397  0.4246680  1.21031392  0.34658271
#> 40  1.81924749  1.02694115 -0.4277478 -0.2110915  0.11341310 -0.56079799
#>            V7
#> 21  1.0887463
#> 22 -0.3394119
#> 23 -0.2276245
#> 24  0.9890037
#> 25  1.9077749
#> 26  0.3631237
#> 27 -0.3745197
#> 28 -1.1917526
#> 29 -0.2773900
#> 30 -0.9261179
#> 31 -0.2005605
#> 32  0.4956025
#> 33 -0.8323755
#> 34  2.8982838
#> 35  0.2416003
#> 36  1.2796943
#> 37 -2.3639139
#> 38  0.8649598
#> 39 -1.3272231
#> 40  1.0554901
#> 
#> 
#> $Y
#>           11           12           13           14           15           16 
#>  0.256670132 -0.548988603  1.182798606 -0.842691693 -0.722074328 -1.142532205 
#>           17           18           19          110          111          112 
#> -0.809638631  0.803724929  0.292242126  0.863699194 -0.531578688  0.234910268 
#>          113          114          115          116          117          118 
#>  0.102704901 -1.567717366  1.636969264 -0.007012363  0.624138975  0.813676653 
#>          119          120           21           22           23           24 
#> -0.192025725 -0.447275444  0.683159577 -1.257812275  1.760823261 -0.594146232 
#>           25           26           27           28           29          210 
#>  1.443631177  1.301698568 -0.127548392  0.356706789 -1.235062605  0.112492377 
#>          211          212          213          214          215          216 
#>  0.833098333 -0.111435639 -0.664651035 -0.866295978 -0.246436548  0.858647335 
#>          217          218          219          220 
#> -0.694489056 -0.331524237 -1.504654130  0.283798710 
#> 
#> $X
#>                V1         V2         V3           V4          V5          V6
#>  [1,] -0.47876585  0.7212135 -0.1765883  2.052937211 -0.44724339 -0.78208645
#>  [2,]  0.26714948  0.5764173 -0.2796235 -0.206029892 -0.01580652 -0.03067421
#>  [3,] -0.67136830 -0.1722877  0.8225721  0.468626546  1.29123874 -1.11164466
#>  [4,]  1.56694673 -2.3561876  0.6598507 -0.143798560 -1.39761854  0.10079849
#>  [5,]  0.40145748  0.4046711 -0.7845974 -0.857599138  0.70986474 -0.82089774
#>  [6,] -0.65740917 -0.3105792 -0.8063316  0.005009635  0.45010515  1.92607908
#>  [7,]  0.54686708 -0.4160399  0.4370732 -1.840168657  1.17679622  0.73447257
#>  [8,]  0.77788495 -1.8074402  0.9055791  1.186628929 -0.18401133  0.95390972
#>  [9,]  0.62821947 -0.7571342 -0.1160725 -0.027880993  0.48699801  0.35731367
#> [10,] -0.18313757  0.1910506  1.0361444  1.840996461  0.38456548  1.82954281
#> [11,]  1.49006255  1.1864772  0.4759545  0.270360939 -0.42121855 -0.79951942
#> [12,]  0.45701273 -0.3599506 -0.6956061 -0.826755844  1.32087849 -0.60205814
#> [13,] -0.47396565  0.1590209  0.4098614  0.395501520  1.27363812  1.54610396
#> [14,] -1.94117948 -0.3081204 -1.2956179 -1.034203486  0.81562886 -0.81648750
#> [15,]  1.13386140 -1.7083006  1.6762616 -1.329927719  1.69804915 -0.31366251
#> [16,]  0.05668208 -0.6903072  2.3111135  0.100013886  0.67457294 -0.52398089
#> [17,]  0.08314817 -0.6683990 -0.4116791 -0.579972159 -1.15181375 -0.44139136
#> [18,]  0.96711434 -0.3139490 -1.1967691 -0.168691040 -0.45181830 -0.39502424
#> [19,]  0.85421378  0.9127865  0.6749937 -0.100912550 -1.10008037  0.48208311
#> [20,]  0.64490384  0.5563545 -0.1424099 -0.715304364 -0.35242654 -0.27957839
#> [21,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [22,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [23,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [24,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [25,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [26,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [27,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [28,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [29,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [30,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [31,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [32,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [33,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [34,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [35,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [36,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [37,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [38,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [39,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#> [40,]  0.00000000  0.0000000  0.0000000  0.000000000  0.00000000  0.00000000
#>                 V7          V1          V2         V3         V4          V5
#>  [1,] -0.463787962  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [2,]  1.506485622  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [3,] -0.153298165  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [4,] -0.115978105  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [5,] -0.031406644  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [6,]  0.834766414  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [7,] -0.003033862  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [8,]  0.035253066  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#>  [9,] -0.651015494  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [10,] -0.270183749  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [11,]  0.139462113  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [12,] -0.552164899  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [13,]  0.641713193  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [14,] -1.542631192  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [15,]  0.402020957  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [16,] -1.561871226  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [17,] -0.245361720  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [18,] -0.487574553  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [19,] -0.619510556  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [20,]  0.014727009  0.00000000  0.00000000  0.0000000  0.0000000  0.00000000
#> [21,]  0.000000000 -1.66463418  0.19862558 -1.4138724  0.2091795 -1.60627712
#> [22,]  0.000000000  1.18142151 -0.50370801  1.1553587  1.3933059  0.12085357
#> [23,]  0.000000000 -1.43500545  0.86882915  0.2692985  1.2986432 -0.50858307
#> [24,]  0.000000000 -0.32875042  0.68676624 -1.6875022 -0.4759629 -0.22062175
#> [25,]  0.000000000 -0.92945305 -0.90642941  0.6185146 -2.2846557 -1.03236239
#> [26,]  0.000000000 -0.59327792  2.08316395 -0.1698973  2.1418239  1.91333305
#> [27,]  0.000000000  2.01986170 -0.52104016  1.7128693  0.4476429 -0.21083180
#> [28,]  0.000000000  0.11407313 -1.75849329 -0.8742812  0.3312741 -1.47917701
#> [29,]  0.000000000 -1.08633470 -0.40398155 -0.4847358 -0.1821150  0.31499123
#> [30,]  0.000000000 -1.41256907 -0.03158415 -1.0598825  0.3023816  0.38066091
#> [31,]  0.000000000  0.51257571  2.19096231 -0.1911791 -0.3218407 -0.86242547
#> [32,]  0.000000000  0.08096633 -0.13923441  0.4804846  0.2196308 -2.75650534
#> [33,]  0.000000000 -0.19481296  0.23237838 -0.8344605 -0.5401347 -0.51871622
#> [34,]  0.000000000 -0.75767558 -0.33282558  0.9773020 -1.4378152  0.68653721
#> [35,]  0.000000000 -1.27155713 -0.60460502 -1.3869177  0.7445114  0.05929142
#> [36,]  0.000000000 -0.89195200 -0.28793224 -1.2012341  1.2367809  0.02101810
#> [37,]  0.000000000  1.01885494  0.58224057  1.6856955 -0.4554475  0.67689618
#> [38,]  0.000000000 -0.47398969  1.94469926 -1.1639614 -1.3296097 -1.06210711
#> [39,]  0.000000000 -1.17668673  0.83593112  0.4920397  0.4246680  1.21031392
#> [40,]  0.000000000  1.81924749  1.02694115 -0.4277478 -0.2110915  0.11341310
#>                V6         V7
#>  [1,]  0.00000000  0.0000000
#>  [2,]  0.00000000  0.0000000
#>  [3,]  0.00000000  0.0000000
#>  [4,]  0.00000000  0.0000000
#>  [5,]  0.00000000  0.0000000
#>  [6,]  0.00000000  0.0000000
#>  [7,]  0.00000000  0.0000000
#>  [8,]  0.00000000  0.0000000
#>  [9,]  0.00000000  0.0000000
#> [10,]  0.00000000  0.0000000
#> [11,]  0.00000000  0.0000000
#> [12,]  0.00000000  0.0000000
#> [13,]  0.00000000  0.0000000
#> [14,]  0.00000000  0.0000000
#> [15,]  0.00000000  0.0000000
#> [16,]  0.00000000  0.0000000
#> [17,]  0.00000000  0.0000000
#> [18,]  0.00000000  0.0000000
#> [19,]  0.00000000  0.0000000
#> [20,]  0.00000000  0.0000000
#> [21,]  0.72381436  1.0887463
#> [22,]  1.09447970 -0.3394119
#> [23,]  0.17503617 -0.2276245
#> [24,] -1.07510631  0.9890037
#> [25,]  1.24063387  1.9077749
#> [26,] -2.34710475  0.3631237
#> [27,] -0.69636543 -0.3745197
#> [28,] -0.36843478 -1.1917526
#> [29,] -0.26686423 -0.2773900
#> [30,]  1.07900436 -0.9261179
#> [31,]  0.07606432 -0.2005605
#> [32,]  0.38338098  0.4956025
#> [33,] -0.15744866 -0.8323755
#> [34,] -0.35935366  2.8982838
#> [35,]  0.71047589  0.2416003
#> [36,]  1.22165106  1.2796943
#> [37,] -2.80412432 -2.3639139
#> [38,]  0.57117882  0.8649598
#> [39,]  0.34658271 -1.3272231
#> [40,] -0.56079799  1.0554901
#> 
#> $var_indicator
#>      [,1] [,2]
#> [1,]    1    8
#> [2,]    2    9
#> [3,]    3   10
#> [4,]    4   11
#> [5,]    5   12
#> [6,]    6   13
#> [7,]    7   14
#> 
#> $fitted_values
#>           4.34242    3.923666    3.715741    2.569188    1.510147       0.938
#>  [1,]  0.63148173  0.63821768  0.63807138  0.61098903  0.57924286  0.59316407
#>  [2,]  0.37042398  0.37259998  0.37073803  0.33134218  0.31976771  0.31207518
#>  [3,]  0.39883622  0.44193019  0.46878247  0.60909179  0.70859997  0.80846822
#>  [4,]  0.31135510  0.26977346  0.25870266  0.31852587  0.40512148  0.45507246
#>  [5,] -0.20873516 -0.18667909 -0.18249836 -0.18327245 -0.10709383 -0.04665221
#>  [6,]  0.07339801  0.07491192  0.06326351  0.04165436 -0.01401002 -0.10650009
#>  [7,] -0.45016513 -0.43374427 -0.42079098 -0.34904714 -0.24507306 -0.21839132
#>  [8,]  0.55502182  0.55226929  0.55680163  0.67651635  0.80269406  0.85268980
#>  [9,] -0.10426536 -0.09153933 -0.08745994 -0.02558853  0.09324218  0.14593760
#> [10,]  0.46191691  0.47523159  0.48672869  0.53677525  0.56837351  0.53401535
#> [11,]  0.24734998  0.24904390  0.25761294  0.21857654  0.30102104  0.36516580
#> [12,] -0.33957735 -0.30244985 -0.29159117 -0.21595991 -0.05635409  0.04532038
#> [13,]  0.21500919  0.24513945  0.25870419  0.32319249  0.36689431  0.34426140
#> [14,] -0.50054642 -0.48819127 -0.49490263 -0.49367737 -0.59167575 -0.62043464
#> [15,]  0.05289330  0.09286815  0.12849251  0.37183710  0.65193192  0.82119940
#> [16,]  0.12392329  0.14091584  0.17284499  0.34179778  0.46142868  0.55263149
#> [17,]  0.05234000  0.01612769 -0.00229886 -0.05818610 -0.13938186 -0.17076839
#> [18,] -0.13941559 -0.14696732 -0.16074497 -0.20557394 -0.14843487 -0.11145052
#> [19,] -0.02939557 -0.06173032 -0.06705751 -0.16101819 -0.21660105 -0.27033795
#> [20,] -0.07875537 -0.08936703 -0.09296107 -0.15000915 -0.15576358 -0.15770668
#> [21,]  0.49221603  0.52399789  0.54673507  0.69495274  0.88694501  0.98118203
#> [22,]  0.46841656  0.47222042  0.46106043  0.39474212  0.31747455  0.33106228
#> [23,]  0.57880526  0.59495558  0.59694516  0.67248962  0.84805671  0.94300663
#> [24,]  0.21840337  0.22936339  0.25080959  0.37578492  0.45516102  0.45652034
#> [25,]  0.23851605  0.23291370  0.22354518  0.15298762  0.11779210  0.10776025
#> [26,]  0.82429394  0.81218211  0.80994497  0.87498836  0.97663585  0.99452225
#> [27,]  0.44526292  0.46150824  0.45505650  0.40776216  0.32143166  0.27258959
#> [28,]  0.15248982  0.18881655  0.21143328  0.25885172  0.29816334  0.27155656
#> [29,]  0.13338265  0.12160985  0.12044642  0.09096421  0.10256322  0.08421978
#> [30,] -0.12104681 -0.13833036 -0.13874728 -0.14747477 -0.11203384 -0.07199898
#> [31,] -0.06349020 -0.04059622 -0.02659010  0.14937523  0.29130187  0.38699472
#> [32,]  0.62531542  0.68823229  0.70927248  0.85748630  1.03205813  1.11463060
#> [33,] -0.18472462 -0.17605177 -0.16350157 -0.09355868 -0.03658196 -0.02289998
#> [34,]  0.72810399  0.70063697  0.67995877  0.55843248  0.46955392  0.40764576
#> [35,]  0.32730393  0.32428758  0.33169044  0.34124291  0.37765775  0.40106830
#> [36,]  0.63636297  0.63814791  0.64453597  0.66916384  0.70250077  0.75317372
#> [37,] -0.13566554 -0.13944002 -0.14882810 -0.18190238 -0.19894358 -0.28039992
#> [38,] -0.14262221 -0.12755702 -0.10926599  0.07289664  0.22353874  0.32009555
#> [39,] -0.05000058 -0.08186691 -0.10030489 -0.16122777 -0.14110122 -0.10943374
#> [40,]  0.22506646  0.23478025  0.24557617  0.32985706  0.29434021  0.28663582
#>          0.817105           0
#>  [1,]  0.59944078  0.82070045
#>  [2,]  0.29736586  0.19209301
#>  [3,]  0.82558403  1.01642139
#>  [4,]  0.46465299  0.52895786
#>  [5,] -0.03265323  0.07122954
#>  [6,] -0.12868439 -0.28694426
#>  [7,] -0.21036663 -0.29426253
#>  [8,]  0.86398922  1.05685500
#>  [9,]  0.16598586  0.35711950
#> [10,]  0.53484698  0.64182248
#> [11,]  0.37990704  0.52180871
#> [12,]  0.07333682  0.31108529
#> [13,]  0.33779100  0.31664422
#> [14,] -0.61709367 -0.65908612
#> [15,]  0.85003949  1.00521343
#> [16,]  0.58132195  0.74017839
#> [17,] -0.17816834 -0.29526182
#> [18,] -0.09611642  0.07630654
#> [19,] -0.27252972 -0.37688082
#> [20,] -0.15698707 -0.21330370
#> [21,]  1.00280704  1.10463045
#> [22,]  0.33386458  0.25982353
#> [23,]  0.96216072  1.00043963
#> [24,]  0.45975855  0.50440208
#> [25,]  0.10911059  0.21476700
#> [26,]  1.00218167  0.93122006
#> [27,]  0.26347551  0.15091376
#> [28,]  0.26300964  0.17956967
#> [29,]  0.07917106  0.07015104
#> [30,] -0.06810332 -0.01978226
#> [31,]  0.40544025  0.55188692
#> [32,]  1.13340877  1.18321115
#> [33,] -0.02340581  0.02563091
#> [34,]  0.40340725  0.41372040
#> [35,]  0.40605949  0.40196361
#> [36,]  0.76726329  0.77272466
#> [37,] -0.30240242 -0.39430653
#> [38,]  0.34043029  0.55571950
#> [39,] -0.10762799 -0.07636306
#> [40,]  0.28902902  0.30624292
#> 
#> $residuals
#>           4.34242      3.923666    3.715741    2.569188     1.510147
#>  [1,] -0.23335158 -0.2400875277 -0.23994123 -0.21285888 -0.181112700
#>  [2,] -0.77795256 -0.7801285641 -0.77826660 -0.73887076 -0.727296289
#>  [3,]  0.92542241  0.8823284429  0.85547616  0.71516684  0.615658663
#>  [4,] -1.01258677 -0.9710051266 -0.95993433 -1.01975754 -1.106353150
#>  [5,] -0.37187914 -0.3939352154 -0.39811594 -0.39734186 -0.473520470
#>  [6,] -1.07447019 -1.0759840976 -1.06433569 -1.04272655 -0.987062162
#>  [7,] -0.21801347 -0.2344343388 -0.24738762 -0.31913147 -0.423105542
#>  [8,]  0.39016313  0.3929156677  0.38838332  0.26866861  0.142490896
#>  [9,]  0.53796751  0.5252414820  0.52116209  0.45929068  0.340459972
#> [10,]  0.54324231  0.5299276233  0.51843053  0.46838396  0.436785704
#> [11,] -0.63746864 -0.6391625644 -0.64773160 -0.60869521 -0.691139704
#> [12,]  0.71594764  0.6788201396  0.66796146  0.59233020  0.432724382
#> [13,]  0.02915573 -0.0009745208 -0.01453926 -0.07902756 -0.122729386
#> [14,] -0.92571092 -0.9380660762 -0.93135471 -0.93257997 -0.834581589
#> [15,]  1.72553598  1.6855611328  1.64993677  1.40659219  1.126497370
#> [16,]  0.01052437 -0.0064681753 -0.03839733 -0.20735012 -0.326981020
#> [17,]  0.71325899  0.7494713080  0.76789786  0.82378510  0.904980861
#> [18,]  1.09455227  1.1021040014  1.11588164  1.16071062  1.103571546
#> [19,] -0.02117013  0.0111646228  0.01649180  0.11045249  0.166035347
#> [20,] -0.22706005 -0.2164483911 -0.21285435 -0.15580627 -0.150051835
#> [21,]  0.40145767  0.3696758123  0.34693863  0.19872096  0.006728688
#> [22,] -1.51571471 -1.5195185675 -1.50835858 -1.44204027 -1.364772699
#> [23,]  1.39253212  1.3763818094  1.37439223  1.29884777  1.123280673
#> [24,] -0.60203547 -0.6129954947 -0.63444170 -0.75941702 -0.838793126
#> [25,]  1.41562926  1.4212315995  1.43060012  1.50115768  1.536353206
#> [26,]  0.68791875  0.7000305812  0.70226772  0.63722433  0.535576846
#> [27,] -0.36229719 -0.3785425024 -0.37209077 -0.32479642 -0.238465926
#> [28,]  0.41473109  0.3784043686  0.35578764  0.30836920  0.269057578
#> [29,] -1.15793113 -1.1461583259 -1.14499490 -1.11551269 -1.127111698
#> [30,]  0.44405331  0.4613368596  0.46175378  0.47048127  0.435040344
#> [31,]  1.10710266  1.0842086739  1.07020255  0.89423723  0.752310587
#> [32,] -0.52623693 -0.5891538038 -0.61019400 -0.75840781 -0.932979643
#> [33,] -0.26941229 -0.2780851431 -0.29063534 -0.36057823 -0.417554950
#> [34,] -1.38388585 -1.3564188200 -1.33574062 -1.21421433 -1.125335773
#> [35,] -0.36322635 -0.3602100028 -0.36761286 -0.37716533 -0.413580172
#> [36,]  0.43279849  0.4310135475  0.42462549  0.39999762  0.366660693
#> [37,] -0.34830939 -0.3445349090 -0.33514683 -0.30207255 -0.285031347
#> [38,]  0.02161210  0.0065469049 -0.01174412 -0.19390675 -0.344548854
#> [39,] -1.24413942 -1.2122730894 -1.19383512 -1.13291223 -1.153038783
#> [40,]  0.26924638  0.2595325813  0.24873667  0.16445578  0.199972624
#>             0.938    0.817105           0
#>  [1,] -0.19503392 -0.20131063 -0.42257030
#>  [2,] -0.71960376 -0.70489444 -0.59962159
#>  [3,]  0.51579041  0.49867460  0.30783724
#>  [4,] -1.15630413 -1.16588466 -1.23018953
#>  [5,] -0.53396209 -0.54796107 -0.65184385
#>  [6,] -0.89457209 -0.87238779 -0.71412792
#>  [7,] -0.44978729 -0.45781197 -0.37391608
#>  [8,]  0.09249516  0.08119573 -0.11167004
#>  [9,]  0.28776455  0.26771629  0.07658265
#> [10,]  0.47114387  0.47031223  0.36333674
#> [11,] -0.75528447 -0.77002570 -0.91192738
#> [12,]  0.33104991  0.30303347  0.06528500
#> [13,] -0.10009648 -0.09362608 -0.07247930
#> [14,] -0.80582270 -0.80916367 -0.76717122
#> [15,]  0.95722989  0.92838979  0.77321585
#> [16,] -0.41818382 -0.44687429 -0.60573073
#> [17,]  0.93636739  0.94376734  1.06086082
#> [18,]  1.06658720  1.05125309  0.87883014
#> [19,]  0.21977225  0.22196401  0.32631512
#> [20,] -0.14810874 -0.14882835 -0.09251172
#> [21,] -0.08750833 -0.10913334 -0.21095675
#> [22,] -1.37836043 -1.38116273 -1.30712168
#> [23,]  1.02833076  1.00917667  0.97089775
#> [24,] -0.84015244 -0.84339066 -0.88803419
#> [25,]  1.54638505  1.54503471  1.43937830
#> [26,]  0.51769045  0.51003102  0.58099264
#> [27,] -0.18962385 -0.18050978 -0.06794802
#> [28,]  0.29566436  0.30421127  0.38765124
#> [29,] -1.10876826 -1.10371954 -1.09469952
#> [30,]  0.39500548  0.39110982  0.34278877
#> [31,]  0.65661774  0.63817221  0.49172553
#> [32,] -1.01555211 -1.03433028 -1.08413266
#> [33,] -0.43123693 -0.43073110 -0.47976782
#> [34,] -1.06342761 -1.05918910 -1.06950226
#> [35,] -0.43699073 -0.44198192 -0.43788603
#> [36,]  0.31598774  0.30189817  0.29643680
#> [37,] -0.20357501 -0.18157251 -0.08966840
#> [38,] -0.44110566 -0.46144040 -0.67672961
#> [39,] -1.18470627 -1.18651202 -1.21777694
#> [40,]  0.20767701  0.20528381  0.18806992
#> 
#> $coef_df
#> # A tibble: 112 × 10
#> # Groups:   knot [8]
#>    class x_var present scale_factor lambda  scaled_coef `rev(as.numeric(l…  knot
#>    <chr> <chr> <lgl>          <dbl> <chr>         <dbl>              <dbl> <int>
#>  1 1     V1    TRUE           1.09  4.34242     -0.0639                  0     1
#>  2 1     V2    TRUE           0.945 4.34242     -0.0655                  0     1
#>  3 1     V3    TRUE           0.862 4.34242      0.127                   0     1
#>  4 1     V4    TRUE           1.08  4.34242      0.257                   0     1
#>  5 1     V5    TRUE           1.00  4.34242     -0.0768                  0     1
#>  6 1     V6    TRUE           0.882 4.34242     -0.103                   0     1
#>  7 1     V7    TRUE           0.938 4.34242      0.244                   0     1
#>  8 2     V1    TRUE           1.09  4.34242     -0.0639                  0     1
#>  9 2     V2    TRUE           0.945 4.34242     -0.0655                  0     1
#> 10 2     V3    TRUE           0.862 4.34242      0.127                   0     1
#> # … with 102 more rows, and 2 more variables: coef <dbl>, df <int>
#> 
#> $raw_data
#>    class          V1          V2         V3           V4           V5
#> 1      1 -0.62645381  0.91897737 -0.1645236  2.401617761 -0.568668733
#> 2      1  0.18364332  0.78213630 -0.2533617 -0.039240003 -0.135178615
#> 3      1 -0.83562861  0.07456498  0.6969634  0.689739362  1.178086997
#> 4      1  1.59528080 -1.98935170  0.5566632  0.028002159 -1.523566800
#> 5      1  0.32950777  0.61982575 -0.6887557 -0.743273209  0.593946188
#> 6      1 -0.82046838 -0.05612874 -0.7074952  0.188792300  0.332950371
#> 7      1  0.48742905 -0.15579551  0.3645820 -1.804958629  1.063099837
#> 8      1  0.73832471 -1.47075238  0.7685329  1.465554862 -0.304183924
#> 9      1  0.57578135 -0.47815006 -0.1123462  0.153253338  0.370018810
#> 10     1 -0.30538839  0.41794156  0.8811077  2.172611670  0.267098791
#> 11     1  1.51178117  1.35867955  0.3981059  0.475509529 -0.542520031
#> 12     1  0.38984324 -0.10278773 -0.6120264 -0.709946431  1.207867806
#> 13     1 -0.62124058  0.38767161  0.3411197  0.610726353  1.160402616
#> 14     1 -2.21469989 -0.05380504 -1.1293631 -0.934097632  0.700213650
#> 15     1  1.12493092 -1.37705956  1.4330237 -1.253633400  1.586833455
#> 16     1 -0.04493361 -0.41499456  1.9803999  0.291446236  0.558486426
#> 17     1 -0.01619026 -0.39428995 -0.3672215 -0.443291873 -1.276592208
#> 18     1  0.94383621 -0.05931340 -1.0441346  0.001105352 -0.573265414
#> 19     1  0.82122120  1.10002537  0.5697196  0.074341324 -1.224612615
#> 20     1  0.59390132  0.76317575 -0.1350546 -0.589520946 -0.473400636
#> 21     2 -1.91435943  0.42510038 -1.2313234  0.409401840 -1.733218407
#> 22     2  1.17658331 -0.23864710  0.9838956  1.688873286  0.002131860
#> 23     2 -1.66497244  1.05848305  0.2199248  1.586588433 -0.630300334
#> 24     2 -0.46353040  0.88642265 -1.4672500 -0.330907801 -0.340968580
#> 25     2 -1.11592011 -0.61924305  0.5210227 -2.285235535 -1.156572363
#> 26     2 -0.75081900  2.20610246 -0.1587546  2.497661590  1.803141908
#> 27     2  2.08716655 -0.25502703  1.4645873  0.667066167 -0.331132036
#> 28     2  0.01739562 -1.42449465 -0.7660820  0.541327336 -1.605513412
#> 29     2 -1.28630053 -0.14439960 -0.4302118 -0.013399523  0.197193439
#> 30     2 -1.64060553  0.20753834 -0.9261095  0.510108423  0.263175646
#> 31     2  0.45018710  2.30797840 -0.1771040 -0.164375832 -0.985826700
#> 32     2 -0.01855983  0.10580237  0.4020118  0.420694643 -2.888920672
#> 33     2 -0.31806837  0.45699881 -0.7317482 -0.400246744 -0.640481703
#> 34     2 -0.92936215 -0.07715294  0.8303732 -1.370207878  0.570507636
#> 35     2 -1.48746031 -0.33400084 -1.2080828  0.987838267 -0.059723276
#> 36     2 -1.07519230 -0.03472603 -1.0479844  1.519745025 -0.098178744
#> 37     2  1.00002880  0.78763961  1.4411577 -0.308740569  0.560820729
#> 38     2 -0.62126669  2.07524501 -1.0158475 -1.253289756 -1.186458639
#> 39     2 -1.38442685  1.02739244  0.4119747  0.642241306  1.096777044
#> 40     2  1.86929062  1.20790840 -0.3810761 -0.044709137 -0.005344028
#>             V6          V7           y
#> 1  -0.62036668 -0.50595746  0.39813016
#> 2   0.04211587  1.34303883 -0.40752858
#> 3  -0.91092165 -0.21457941  1.32425863
#> 4   0.15802877 -0.17955653 -0.70123167
#> 5  -0.65458464 -0.10019074 -0.58061430
#> 6   1.76728727  0.71266631 -1.00107218
#> 7   0.71670748 -0.07356440 -0.66817861
#> 8   0.91017423 -0.03763417  0.94518495
#> 9   0.38418536 -0.68166048  0.43370215
#> 10  1.68217608 -0.32427027  1.00515922
#> 11 -0.63573645  0.06016044 -0.39011866
#> 12 -0.46164473 -0.58889449  0.37637029
#> 13  1.43228224  0.53149619  0.24416492
#> 14 -0.65069635 -1.51839408 -1.42625734
#> 15 -0.20738074  0.30655786  1.77842929
#> 16 -0.39280793 -1.53644982  0.13444766
#> 17 -0.31999287 -0.30097613  0.76559900
#> 18 -0.27911330 -0.52827990  0.95513668
#> 19  0.49418833 -0.65209478 -0.05056570
#> 20 -0.17733048 -0.05689678 -0.30581542
#> 21  0.70731067  0.95101281  0.89367370
#> 22  1.03410773 -0.38923718 -1.04729815
#> 23  0.22348041 -0.28433066  1.97133739
#> 24 -0.87870761  0.85740978 -0.38363211
#> 25  1.16296456  1.71962730  1.65414530
#> 26 -2.00016494  0.27005490  1.51221269
#> 27 -0.54479074 -0.42218401  0.08296573
#> 28 -0.25567071 -1.18911329  0.56722091
#> 29 -0.16612104 -0.33103298 -1.02454848
#> 30  1.02046391 -0.93982933  0.32300650
#> 31  0.13622189 -0.25893258  1.04361246
#> 32  0.40716760  0.39437917  0.09907849
#> 33 -0.06965481 -0.85185709 -0.45413691
#> 34 -0.24766434  2.64916688 -0.65578185
#> 35  0.69555081  0.15601168 -0.03592242
#> 36  1.14622836  1.13020727  1.06916146
#> 37 -2.40309621 -2.28912398 -0.48397493
#> 38  0.57273956  0.74100116 -0.12101011
#> 39  0.37472441 -1.31624516 -1.29414000
#> 40 -0.42526772  0.91980368  0.49431284
#> 
#> $class_var
#> [1] "class"
#> 
#> $y_var
#> [1] "y"
#> 
#> $scale
#> [1] TRUE
#> 
#> $center_y
#> [1] TRUE
#> 
#> $class_means_vector
#>         1         1         1         1         1         1         1         1 
#> 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 
#>         1         1         1         1         1         1         1         1 
#> 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 0.1414600 
#>         1         1         1         1         2         2         2         2 
#> 0.1414600 0.1414600 0.1414600 0.1414600 0.2105141 0.2105141 0.2105141 0.2105141 
#>         2         2         2         2         2         2         2         2 
#> 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 
#>         2         2         2         2         2         2         2         2 
#> 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 0.2105141 
#> 
#> $alpha
#> # A tibble: 16 × 3
#> # Groups:   class [2]
#>    class  knot   alpha
#>    <chr> <int>   <dbl>
#>  1 1         1  0.187 
#>  2 1         2  0.183 
#>  3 1         3  0.180 
#>  4 1         4  0.165 
#>  5 1         5  0.137 
#>  6 1         6  0.126 
#>  7 1         7  0.120 
#>  8 1         8  0.0716
#>  9 2         1  0.137 
#> 10 2         2  0.127 
#> 11 2         3  0.122 
#> 12 2         4  0.0636
#> 13 2         5 -0.0110
#> 14 2         6 -0.0503
#> 15 2         7 -0.0586
#> 16 2         8 -0.0940
#> 
#> $predictions
#> # A tibble: 320 × 8
#> # Groups:   obs_no, class, knot [320]
#>    obs_no class  knot    df prediction      y  alpha  yhat
#>     <int> <chr> <int> <int>      <dbl>  <dbl>  <dbl> <dbl>
#>  1      1 1         1     7      0.504  0.398 0.187  0.691
#>  2      1 1         2     8      0.508  0.398 0.183  0.691
#>  3      1 1         3     9      0.507  0.398 0.180  0.687
#>  4      1 1         4    10      0.460  0.398 0.165  0.626
#>  5      1 1         5    11      0.405  0.398 0.137  0.542
#>  6      1 1         6    12      0.403  0.398 0.126  0.528
#>  7      1 1         7    13      0.407  0.398 0.120  0.527
#>  8      1 1         8    14      0.614  0.398 0.0716 0.686
#>  9      2 1         1     7      0.243 -0.408 0.187  0.430
#> 10      2 1         2     8      0.243 -0.408 0.183  0.426
#> # … with 310 more rows
#> 
#> $scaled_predictions
#> # A tibble: 320 × 5
#> # Groups:   obs_no, class [40]
#>    obs_no class  knot scaled_prediction      y
#>     <int> <chr> <int>             <dbl>  <dbl>
#>  1      1 1         1             0.490  0.257
#>  2      1 1         2             0.497  0.257
#>  3      1 1         3             0.497  0.257
#>  4      1 1         4             0.470  0.257
#>  5      1 1         5             0.438  0.257
#>  6      1 1         6             0.452  0.257
#>  7      1 1         7             0.458  0.257
#>  8      1 1         8             0.679  0.257
#>  9      2 1         1             0.229 -0.549
#> 10      2 1         2             0.231 -0.549
#> # … with 310 more rows
#> 
#> $lambdas
#> [1] 4.342420 3.923666 3.715741 2.569188 1.510147 0.938000 0.817105 0.000000
#> 
#> $bic
#> # A tibble: 8 × 8
#> # Groups:   knot [8]
#>    knot    df sigma2     n     aic  bic1  bic2  ebic
#>   <int> <dbl>  <dbl> <int>   <dbl> <dbl> <dbl> <dbl>
#> 1     1     7  0.641    40 -0.0955 0.200  1.29 0.662
#> 2     2     8  0.630    40 -0.0625 0.275  1.37 0.803
#> 3     3     9  0.622    40 -0.0251 0.355  1.45 0.949
#> 4     4    10  0.572    40 -0.0588 0.363  1.49 1.02 
#> 5     5    11  0.526    40 -0.0924 0.372  1.54 1.10 
#> 6     6    12  0.505    40 -0.0833 0.423  1.61 1.22 
#> 7     7    13  0.501    40 -0.0416 0.507  1.70 1.37 
#> 8     8    14  0.477    40 -0.0412 0.550  1.77 1.47 
#> 
#> attr(,"class")
#> [1] "mcreg"